Move yaml helpers to utils
